test(cli): split file command tests into per-feature modules
Break the monolithic index.cli.test.ts into four focused test files
(download, download-resume, list-cleanup, upload) for easier navigation
and independent execution.
